pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coxs-bazar-itinerary-mcp-server"
version = "0.1.0"
description = "Cox's Bazar AI Itinerary MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
cox-mcp-server = "mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
]
